% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BACKCHANNEL_CUES}
\alias{BACKCHANNEL_CUES}
\title{Backchannel Cue Words}
\format{
A character vector of 31 words
}
\source{
Cooney & Reece (2025). Scientific Reports.
  \url{https://doi.org/10.1038/s41598-025-24381-1}
}
\usage{
BACKCHANNEL_CUES
}
\description{
A character vector of words that are typically used as backchannel cues
(e.g., "yeah", "okay", "mhm", "right"). Used by the backchannel detection
algorithm to classify overlapping speech.
}
\keyword{datasets}
