% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOT_BACKCHANNEL_CUES}
\alias{NOT_BACKCHANNEL_CUES}
\title{Not Backchannel Cue Words}
\format{
A character vector of 15 words
}
\source{
Cooney & Reece (2025). Scientific Reports.
  \url{https://doi.org/10.1038/s41598-025-24381-1}
}
\usage{
NOT_BACKCHANNEL_CUES
}
\description{
A character vector of words that, when used at the start of an utterance,
indicate it is NOT a backchannel but rather a secondary turn (e.g., "and",
"but", "i", "well"). Used by the backchannel detection algorithm.
}
\keyword{datasets}
